all: fix and enable "-Wsign-compare" warning

+++ lib/addr.c
@@ -1042,8 +1042,8 @@ char *nl_addr2str(const struct nl_addr *addr, char *buf, size_t size)
 
 prefix:
 	if (addr->a_family != AF_MPLS &&
-	    addr->a_prefixlen != (8 * addr->a_len)) {
-		snprintf(tmp, sizeof(tmp), "/%u", addr->a_prefixlen);
+	    (unsigned)addr->a_prefixlen != (8u * ((size_t)addr->a_len))) {
+		snprintf(tmp, sizeof(tmp), "/%d", addr->a_prefixlen);
 		strncat(buf, tmp, size - strlen(buf) - 1);
 	}
